Fix setup.py yaml dependency issue #20.
Because we prepare the adapters when the module is loaded the import yaml dependency comes into play before setup has a chance to install the dependencies.
Setup retrieves the version information from __version__ which effectively loads the aero module.
Moved the yaml dependencies to the function call where they are required and therefor does not affect setup.py any longer.

setup.py install

Now collects and installs all the dependencies or fails if this is not possible.
PyYaml requires lbyaml to be installed or it will fail to build.
Checking for the presence of PyYaml's dependencies is outside the scope of this project.

This commit resolves #20.