Coroutine autostart annotation wrapper.
Coroutines as apposed to Generators that produce information sits and waits to consume information.
But what makes them awkward is that you need to start each one first by calling next (as it's still just an iterator) to get it going.
This wrapper allows you to mark each coroutine with annotation which serves as perfect informative measures and facilitates the starting process for us.