Wiring the coroutines into process
Like unix pipes coroutines no deliver a very flexible way to create a process as each routine receives a target, the next routine it forwards to. It is completely independent of the caller and can be called from multiple sources.

For our simple routine we don't need to keep many references as each routine holds its own target and this can now be overwritten should a subclass desire a slightly different approach.