Enable pass throuch variables.
It's understandable that you want to enable some special flags generic enough to be part of aeros functionality. This will now allow you to add command arguments to be added at command execution.

I am still not completely hoppy with the way argparse deals with these, it does something funny with single - commands even in quotes. This needs to be looked at some more.